fix:ranamign: renamted few app/providers
