+++ backend/src/app/providers/jwt.provider.interface.ts
